fix: bug with boardOwner
